requirements.txt
torch>=2.0.0
transformers>=4.30.0
gradio>=4.0.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
google-api-python-client>=2.0.0
python-docx>=0.8.11
Pillow>=9.0.0
requests>=2.26.0
beautifulsoup4>=4.9.3
safetensors>=0.3.1 